+++ src/OpenTelemetry.Instrumentation.AWSLambda/Implementation/AWSLambdaResourceDetector.cs
@@ -26,9 +26,8 @@ internal sealed class AWSLambdaResourceDetector : IResourceDetector
                 .AttributeBuilder
                 .AddAttributeCloudProviderIsAWS()
                 .AddAttributeCloudRegion(AWSLambdaUtils.GetAWSRegion())
-                .AddAttributeFaasName(AWSLambdaUtils.GetFunctionName())
-                .AddAttributeFaasVersion(AWSLambdaUtils.GetFunctionVersion())
                 .AddAttributeFaasInstance(AWSLambdaUtils.GetFunctionInstance())
+                .AddAttributeFaasInstanceId(AWSLambdaUtils.GetFunctionInstance())
                 .AddAttributeFaasMaxMemory(AWSLambdaUtils.GetFunctionMemorySize())
                 .Build();